+++ lib/infopark_kickstarter.rb
@@ -1,0 +1,11 @@
+require 'jquery-rails'
+require 'less-rails-bootstrap'
+require 'haml'
+require 'infopark_kickstarter/engine'
+require 'infopark_kickstarter/configuration'
+require 'infopark_kickstarter/dashboard'
+require 'generators/cms/migration'
+require 'generators/cms/base_paths'
+
+module InfoparkKickstarter
+end