+++ lib/infopark_kickstarter/rake/credential_helper.rb
@@ -1,0 +1,17 @@
+module InfoparkKickstarter
+  module Rake
+    module CredentialHelper
+      def url
+        config['url']
+      end
+
+      def api_key
+        config['api_key']
+      end
+
+      def config
+        YAML.load_file(Rails.root + 'config/deploy.yml')
+      end
+    end
+  end
+end