+++ template.rb
@@ -6,7 +6,7 @@ gem_group(:assets) do
 end
 
 gem_group(:test, :development) do
-  gem('ice_kickstarter', path: '../../../')
+  gem('infopark_kickstarter', path: '../../../')
 end
 
 run('bundle --quiet')